django_restapi/requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
